Initial implementation of tar creation leveraging archive/tar.
Added a noodlesCondensedName variable so we place more nicely with file names. Right now we're handling only Go binary creation for the Go plugin as part of tar, and we're doing all this in a unique temporary directory that gets cleaned up after a successful or failed tar creation. Right now error handling is a bit verbose, expected to clean that up later.