Initial commit of plugin-oriented re-architecture.
This is quite a significant commit of the re-architecture of Noodles. Previously, plugin-related (as in Go, LESS, TypeScript) builds were implemented in NoodlesProject, which resulted in an unnecessarily complex build process, as well as eliminating any form of consistency in the plugins supported.

This extended to configuration and linting as well, with half of the error handling for configs happening with linting, the other half in config, as well as plugin-related linting of projects.

With the new architecture, we're enforcing consistency of pre-run, run, and post-run operations, which are implemented in types which implement NoodlesPlugin. Pre-run operations are primarily utilized for binary checking, or in the case of Go it's used for our Go environment setting as well. Run tasks will handle compiling of a specific plugin type as it relates to the passed project, and post-run performs any necessary cleanup, resetting of the environment, and in the case of TypeScript it'll check if it should do compression.

This dramatically simplifies our build process while making it more flexible as well. We're able to gracefully fail more often and avoid run or post-run operations entirely.

This commit also dropkicks init in favor of our upcoming "new" (current WIP), which handles creating the Noodles workspace, a project, or script. This is done in a far more interactive manner, leveraging manifoldco/promptui to make creation more straight-forward and proactively address issues.

Other things to more-or-less ignore in this commit: Early vendoring code that isn't functional yet, even the current code is gonna get rewritten after the "new" bootstrap functionality is finished.