Dropped use of buggy promptui.Prompt, ensure project creation works.
promptui.Prompt wouldn't reliably show characters without a terminal resize, so dropped it in favor of a wrapper around coreutils.InputMessage with a validate func option, so we don't lose our on any functionality that Prompt would've given us. As a result, implemented a common TextYNValidate function for common y/N or yes / no responses to some queries (like compression).

Always read the config when we start noodles, even if it is for workspace creation. It should gracefully fail.