Multiple fixes to compilation and linting.
We will no longer include Go test files during compilation by passing _test.go to our updated NoodlesProject.GetFiles() func.

We are performing a more refined setting of a NoodlesProject destination in the Go plugin now. Previously we'd always assume n.SimpleName was set and we were building a binary. Now we'll check if a destination is set and if it isn't and its a binary, we'll create a build folder specific destination and then attempt to create the directory, otherwise we just set it to the workdir.

During project creation, we'll now enforce lowercased plugin names. During linting, if we do not find a matching plugin then we'll notify the user and immediately exit, rather than continuing and then segfaulting.