Reduce unnecessary prompting during project creation.
Our project creation will no longer prompt for the source and destinations during Go-plugin type projects which are not binaries, since we'll be referring to it as the respective package name instead.

As such, moved this prompting to SourceDestinationPrompt func. Performed further refinements to arg setting and fixed failing to catch compilation failures in Go plugin.