Ensure we move to our Go dir during UseGoEnv in scripts before dir change
If we have specified a directory to switch to in a script, while using the Go env, ensure we switch to the faux Go workspace first before we attempt directory changes. Set our script.Directory value to either point to its value respective to the Go env dir or the workdir based on if UseGoEnv is set.