Add more of our Golang project features to project creation
We should prompt the user to enable or disable features for:

- Enabling Go Modules (EnableGoModules)
- Enabling nested directories and consolidating them during compilation (ConsolidateChildDirs)
- Enabling / disabling our self-contained Go workspace (overrides GOPATH temporarily, DisableNestedEnvironment since enabled is default)

Cleanup less system NPM dependencies and remove --math=strict since that has been integrated for several releases now and the math option, set to always currently, is more sane.

For Go projects, ensure we're not trying to consolidate hidden directories.