Rewrite our file consolidation functionality
Rewrite our file consolidation functionality to not assume our source or target directory is actually the Noodles project SourceDir. Also eliminated redundant ConsolidateFiles code which handled basic checking of "is this file actually already in the target directory or should be excluded".

This rewrite is in preparation for the implementation of support for Go Module file consolidation when those modules are actually Noodles workspaces.