Implemented file consolidation for nested Noodles workspaces
We will now perform file consolidation, passing in our parent Noodle Workspace ExcludeItems for the project, for all Noodles Workspaces found in our Go Modules, if both EnableGoModules is set and DisableNestedEnvironment is NOT set.

Cleaned up ReadConfig to take in a conf, not immediately set the noodles NoodlesConfig variable but rather return a NoodlesConfig struct for us to assign. This may prove useful for future, more elaborate vendor support.

Ensure we clean up any prefixed / in the leading path in Flatten to eliminate unnecessary __ at the beginning of the file. Simplified removal by passing the full file path of what we'll be removing, rather than just the file name. This allows us to perform cleanup of nested Noodles Workspaces.

Ensure that, when required, we will set the Noodles project SimpleName (used in our tracking temp files map).